Creating Typescript functions for main panel. For interactions between Cards and Virtualmachines in server side.

+++ applications/web/app_hypervisor/views.py
@@ -11,6 +11,7 @@ from django.shortcuts import redirect
 
 # Own libraries and functions
 from utils.validations import no_cache_render
+from app_tlk.tlk.utilities.hypervisor import Hypervisor
 
 # Custom form
 from .forms import CreateVirtualMachineForm
@@ -20,6 +21,7 @@ from .models import VirtualMachine, Resources, Services
 
 
 ## app_hypervisor application views
+hypervisor = Hypervisor()
 
 @login_required
 def monitor(request):
@@ -43,7 +45,10 @@ def create_virtual_machine(request):
             application_name = form.cleaned_data['applicationName']
             version = form.cleaned_data['version']
             hardware = form.cleaned_data['hardware']
-            
+
+            number_service_chosen = hardware.split(":")[1]
+            number_service_chosen = number_service_chosen.strip()
+
             chosen_service = Services.objects.get(choice=hardware)
             chosen_resources = chosen_service.description.split("|")
             chosen_vm_resources = [ resource.strip() for resource in chosen_resources ]
@@ -66,19 +71,29 @@ def create_virtual_machine(request):
                     vdisk=chosen_vm_resources[3]
                 )
 
-                print(new_vm_resources)
+                #Sprint(new_vm_resources)
 
                 vm_user =  User.objects.get(username=user.username)
                 VirtualMachine.objects.create(
                     id = uuid.uuid1(),
                     name = virtual_machine_name,
                     app_name = application_name,
+                    app_version = version,
                     user = vm_user,
                     resources = new_vm_resources,
                     creation_date = timezone.now()
                 )
 
                 chosen_service.users.add(vm_user)
+
+                ''' 
+                Create new virtual machine for this User
+                this take three parameters:
+                    1. Name for virtual machine
+                    2. Operating system name
+                    3. Option choice, is a number
+                '''
+                hypervisor.createNewVirtualMachine(virtual_machine_name, chosen_vm_resources[0], number_service_chosen)
 
                 # Redirige al usuario a alguna página de éxito
                 messages.success(request, '&#128516; New virtual machine has been created successfully! ')
@@ -94,11 +109,34 @@ def create_virtual_machine(request):
 
 
 @login_required
+def start(request, vm_uuid):
+    virtual_machine = VirtualMachine.objects.get(id=vm_uuid)
+    virtual_machine.state = True
+    virtual_machine.save()
+    print(f'START -> {virtual_machine}')
+    hypervisor.startVM(virtual_machine.name)
+
+    return JsonResponse({'status':f'{vm_uuid}'})
+#End_def
+
+
+@login_required
+def shutdown(request, vm_uuid):
+    virtual_machine = VirtualMachine.objects.get(id=vm_uuid)
+    virtual_machine.state = False
+    virtual_machine.save()
+    hypervisor.shutdownVM(virtual_machine.name)
+    
+    return JsonResponse({'status':f'{vm_uuid}'})
+#End_def
+
+
+@login_required
 def delete(request, vm_uuid=None):
     if vm_uuid is None and request.method=='GET':
         
         user = request.user
-        username = user.username
+        #username = user.username
 
         if user.is_authenticated:
             context = {
@@ -109,9 +147,24 @@ def delete(request, vm_uuid=None):
     elif request.method=='DELETE':
         try:
             virtual_machine = VirtualMachine.objects.get(id=uuid.UUID(vm_uuid))
-            virtual_machine.delete()
-            print(f'Deleted >> {vm_uuid}')
-            return JsonResponse({'status':f'{vm_uuid}', 'message':'Virtual Machine deleted successfully'})
+            
+            if virtual_machine.name in hypervisor.getNamesOfRunningVM():
+                return JsonResponse(
+                    {
+                        'status':'error', 
+                        'message':f"The virtual Machine called '{virtual_machine.name}' cannot be delete. Because it's running. Please, shutdown first!"
+                    }
+                )
+            else:
+                # Remove from database
+                virtual_machine.delete()
+
+                # Remove from Hypervisor register
+                hypervisor.deleteVM(virtual_machine.name)
+
+                print(f'Deleted >> {vm_uuid}')
+                return JsonResponse({'status':f'{vm_uuid}', 'message':'Virtual Machine deleted successfully'})
+        
         except VirtualMachine.DoesNotExist:
             return JsonResponse({'status': 'error', 'message': 'Error'})
 #End_def
